src/requirements.txt
cloudshell-shell-core>=3.1,<3.2
cloudshell-automation-api>=7.0,<7.1
cloudshell-traffic>=1.0.2
pyxenamanager>=0.6.0
